+++ packages/common-state/package.json
@@ -32,8 +32,9 @@
   },
   "devDependencies": {
     "@pragma-web-utils/core": "^0.4.0",
+    "@testing-library/dom": "^10.4.0",
     "@testing-library/react": "16.0.0",
-    "@testing-library/react-hooks": "^8.0.1",
+    "jest-environment-jsdom": "^29.7.0",
     "ts-jest": "29.2.3"
   },
   "description": "types",
