+++ packages/common-state/src/hooks/HookCommonState.spec.ts
@@ -1,5 +1,5 @@
 import { Destructor } from '@pragma-web-utils/hooks'
-import { act, renderHook } from '@testing-library/react-hooks'
+import { act, renderHook } from '@testing-library/react'
 import { useEffect } from 'react'
 import { TestIncrementor } from '../testUtils'
 import { CommonState } from '../types'
