+++ packages/common-state/src/hooks/useMapCommonState.spec.ts
@@ -1,5 +1,5 @@
 import { Deps } from '@pragma-web-utils/hooks'
-import { act, renderHook } from '@testing-library/react-hooks'
+import { act, renderHook } from '@testing-library/react'
 import { useEffect } from 'react'
 import { TestIncrementor } from '../testUtils'
 import { useCommonState } from './useCommonState'
