+++ packages/common-state/src/hooks/useStrictStateValueContext.spec.ts
@@ -1,4 +1,4 @@
-import { renderHook } from '@testing-library/react-hooks'
+import { renderHook } from '@testing-library/react'
 import React from 'react'
 import { NOT_PROVIDED_STATE_STRICT_CONTEXT } from '../constants'
 import { CommonState } from '../types'
