+++ packages/hooks/package.json
@@ -25,8 +25,9 @@
   },
   "devDependencies": {
     "@pragma-web-utils/core": "^0.4.0",
+    "@testing-library/dom": "^10.4.0",
     "@testing-library/react": "16.0.0",
-    "@testing-library/react-hooks": "^8.0.1",
+    "jest-environment-jsdom": "^29.7.0",
     "ts-jest": "29.2.3"
   },
   "peerDependencies": {
