+++ packages/hooks/src/hooks/useCancelablePool.spec.ts
@@ -1,5 +1,5 @@
 import { wait } from '@pragma-web-utils/core'
-import { renderHook } from '@testing-library/react-hooks'
+import { renderHook } from '@testing-library/react'
 import React from 'react'
 import { CancellablePromise } from '../types'
 import { useCancelablePool } from './useCancelablePool'
