+++ packages/hooks/src/hooks/useDebounce.spec.ts
@@ -1,4 +1,4 @@
-import { renderHook } from '@testing-library/react-hooks'
+import { renderHook } from '@testing-library/react'
 import { useDebounce } from './useDebounce'
 
 describe('useDebounce hook', () => {
