+++ packages/hooks/src/hooks/useStrictContext.spec.ts
@@ -1,4 +1,4 @@
-import { renderHook } from '@testing-library/react-hooks'
+import { renderHook } from '@testing-library/react'
 import React from 'react'
 import { NOT_PROVIDED_STRICT_CONTEXT } from '../constants'
 import { useStrictContext } from './useStrictContext'
