+++ packages/theme-service/src/hooks/useThemeLayoutEffect.spec.ts
@@ -1,4 +1,4 @@
-import { renderHook } from '@testing-library/react-hooks'
+import { renderHook } from '@testing-library/react'
 import { ColorConstant, ThemeConfig } from '../types'
 import { useThemeLayoutEffect } from './index'
